Refactor Treatment-based Card classes
