+++ src/main/java/duke/ui/ImpressionWindow.java
@@ -11,8 +11,12 @@ import duke.data.Plan;
 import duke.data.Result;
 import duke.data.Treatment;
 import duke.ui.card.EvidenceCard;
+import duke.ui.card.InvestigationCard;
+import duke.ui.card.MedicineCard;
 import duke.ui.card.ObservationCard;
+import duke.ui.card.PlanCard;
 import duke.ui.card.ResultCard;
+import duke.ui.card.TreatmentCard;
 import javafx.collections.MapChangeListener;
 import javafx.fxml.FXML;
 import javafx.scene.control.Label;
