--- src/main/java/duke/ui/TreatmentCard.java
@@ -1,6 +1,0 @@
-package duke.ui;
-
-import javafx.scene.layout.AnchorPane;
-
-public class TreatmentCard extends AnchorPane {
-}
